requirements.txt
flask==1.1.1
requests==2.22.0
lxml==4.3.4

# Google services
google-cloud-firestore==1.3.0
google-cloud-tasks==1.1.0
googleapis-common-protos==1.6.0

# Required for requests through HTTPS
ndg-httpsclient==0.5.1
